Day 13-2.1
The displaying now runs separately from the deriving. I had to push everything into a function to get the display to work--for some reason, the previous implementation had this leaking problem where later steps would appear on previous steps--for example, the complete letters from step 12 would appear in steps 11 or 10. Very strange. Also for some reason doing this made the display a lot faster? The previous implementation displayed fold 1 in 3-4 seconds, but this one did it in less than a second. Weird.

Oh, I forgot to mention this previously, but in case you're curious--I used "." instead of " " because with " " I couldn't tell where the paper ended. "." makes it unambiguous.