Day 15-1
Coolio. The solution wasn't having a good time, so I implemented caching and it worked out much better.
Let's just hope that part 2 doesn't involve backtracking, or require multiple paths to be stored...

Also added clearText to handleFiles so that you don't have to reload if you want to see a different day.