Day 19-1
I got sucked into writing a 3d graph in Desmos and didn't complete it until now. Oops! (I'm pretty proud of the 3d graph though.)

The last error was... forgetting that I modified currScan in the for(i) block and assuming it was still the same on subsequent loops.

Speaking of which, I really like my currScan -> reScan converter. It's the sort of thing you could only do in JS.