Day 23-2
I was a little worried, but it seems that caching has saved the day once again! Friendship ended with brute force--now caching is my best friend!

Also, there was array shenanigans--it turns out slice() copies references to objects stored in arrays. So there was a fun bug there where I was just wondering, "Why are all of my arrays changing??"